+++ packages/gp-api/src/campaignStrategy/campaignStrategy.module.ts
@@ -3,6 +3,8 @@ import { Module } from '@nestjs/common'
 import { ClerkModule } from '@/vendors/clerk/clerk.module'
 import { CampaignStrategyController } from './campaignStrategy.controller'
 import { CampaignStrategyService } from './services/campaignStrategy.service'
+import { CommunityEventsPersister } from './services/communityEvents.persister'
+import { CommunityEventsService } from './services/communityEvents.service'
 import { ElectionApiService } from './services/electionApi.service'
 import { StrategicLandscapePersister } from './services/strategicLandscape.persister'
 import { StrategicLandscapeService } from './services/strategicLandscape.service'
@@ -14,6 +16,8 @@ import { StrategicLandscapeService } from './services/strategicLandscape.service
     CampaignStrategyService,
     StrategicLandscapeService,
     StrategicLandscapePersister,
+    CommunityEventsService,
+    CommunityEventsPersister,
     ElectionApiService,
   ],
 })